requirements.txt
flask==3.0.0
anthropic==0.21.0
psycopg2-binary==2.9.9
pandas==2.1.4
openpyxl==3.1.2
python-dotenv==1.0.0
requests==2.31.0
